XmlParser : Changed how empty XML nodes get parsed: The value for an empty node is now saved as an empty dict, i.e. {}, instead of None. This change solves issue #4.
Problem was that None is not iterable, and code such as the following had caused an exception:
if '_attributes' in my_parsed_xml_dict[node_key]:
	...
because my_parsed_xml_dict[node_key] had been None.

Now, with this code change in place, my_parsed_xml_dict[node_key] would resolve not anylonger to None but to an empty dict ( {} ). As the dict is formally iterable (even when empty), above code does not crash.